refactor: ARM-242 move useElementStatus hook to image component, as this is the only place it's being used.
