+++ module/src/components/image/useElementStatus.ts
